+++ src/Warehouses/BrewUp.Warehouse.SharedKernel/Dtos/BeerJson.cs
@@ -4,6 +4,6 @@ public class BeerJson
 {
 	public string BeerId { get; set; } = string.Empty;
 	public string BeerName { get; set; } = string.Empty;
-	public double Stock { get; set; } = 0;
-	public double Availability { get; set; } = 0;
+	public decimal Stock { get; set; } = 0;
+	public decimal Availability { get; set; } = 0;
 }
